backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
# passlib 1.7.4 has a known compat issue with bcrypt>=4.x
# The shim in app/core/security.py handles this transparently.
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
pydantic[email]==2.6.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
gunicorn==21.2.0
